model loading, websocket connection fixes

+++ src/Menubar.js
@@ -1,12 +1,25 @@
 const {Menu, app, shell, dialog} = require('electron');
 
-const fs = require('fs');
+const fs = require('fs-extra');
 const path = require('path');
 
 const {COPYFILE_EXCL} = fs.constants;
 const defaultMenu = require('electron-default-menu');
 
 const menu = defaultMenu(app, shell);
+const uploadFolder = path.join(
+  app.getAppPath(),
+  'src',
+  'student',
+  'public',
+  'uploads',
+);
+
+function containsObj(filePath) {
+  return fs
+    .readdirSync(filePath)
+    .find(file => file.toLowerCase().endsWith('.obj'));
+}
 
 function getMenu(win) {
   const menu = defaultMenu(app, shell);
@@ -21,11 +34,23 @@ function getMenu(win) {
             win,
             {
               filters: [
-                {name: 'Images', extensions: ['jpg', 'png', 'gif']},
-                {name: 'Movies', extensions: ['mkv', 'avi', 'mp4']},
-                {name: '3D models', extensions: ['obj', 'gltf2', 'gltf']},
+                {
+                  name: 'Dateien',
+                  extensions: [
+                    'jpg',
+                    'png',
+                    'gif',
+                    'mkv',
+                    'avi',
+                    'mp4',
+                    'obj',
+                    'gltf2',
+                    'gltf',
+                    'glb',
+                  ],
+                },
               ],
-              properties: ['openFile'],
+              properties: ['openFile', 'openDirectory'],
             },
             filePaths => {
               const basepath = app.getAppPath();
@@ -38,14 +63,55 @@ function getMenu(win) {
                 'uploads',
                 fileName,
               );
+              if (fs.lstatSync(filePaths[0]).isDirectory()) {
+                const objFile = containsObj(filePaths[0]);
+                if (objFile) {
+                  // copy entire folder
+                  fs.copy(filePaths[0], newPath, a => {
+                    console.log(a);
+                    // update menu
+                    Menu.setApplicationMenu(getMenu(win));
 
-              fs.copyFile(filePaths[0], newPath, COPYFILE_EXCL, a => {
-                console.log(a);
-                // update menu
-                Menu.setApplicationMenu(getMenu(win));
+                    win.webContents.send('open', path.join(fileName, objFile));
+                  });
+                } else {
+                  dialog.showErrorBox(
+                    'Ordner enthalt keine .obj-Datei',
+                    'Es konnen nur Ordner, die .obj-Modellen enthalten, geladen werden.',
+                  );
+                }
+              } else if (
+                fileName.endsWith('.obj') &&
+                fs.existsSync(filePaths[0].replace('.obj', '.mtl'))
+              ) {
+                dialog.showErrorBox(
+                  'OBJ Modell besteht aus mehreren Dateien',
+                  'Dieses Modell besteht aus mehreren einzelnen Dateien (.obj, .mtl und ggf. Texturen). Um dieses OBJ-Modell zu laden muss der gesamte Ordner mit allen benötigten Dateien geöffnet werden',
+                );
+              } else if (
+                fileName.endsWith('.gltf') &&
+                fs.existsSync(filePaths[0].replace('.gltf', '.bin'))
+              ) {
+                fs.copyFileSync(
+                  filePaths[0].replace('.gltf', '.bin'),
+                  newPath.replace('.gltf', '.bin'),
+                  COPYFILE_EXCL,
+                );
+                fs.copyFile(filePaths[0], newPath, COPYFILE_EXCL, a => {
+                  console.log(a);
+                  // update menu
+                  Menu.setApplicationMenu(getMenu(win));
+                  win.webContents.send('open', fileName);
+                });
+              } else {
+                fs.copyFile(filePaths[0], newPath, COPYFILE_EXCL, a => {
+                  console.log(a);
+                  // update menu
+                  Menu.setApplicationMenu(getMenu(win));
 
-                win.webContents.send('open', fileName);
-              });
+                  win.webContents.send('open', fileName);
+                });
+              }
             },
           );
         },
@@ -53,13 +119,15 @@ function getMenu(win) {
       {
         label: 'Recent Media',
         submenu: fs
-          .readdirSync(
-            path.join(app.getAppPath(), 'src', 'student', 'public', 'uploads'),
-          )
-          .filter(file => !file.startsWith('.'))
+          .readdirSync(uploadFolder)
+          .filter(file => !file.startsWith('.') && !file.endsWith('.bin'))
           .map(file => ({
             label: file,
             click: () => {
+              if (fs.lstatSync(path.join(uploadFolder, file)).isDirectory()) {
+                const objFile = containsObj(path.join(uploadFolder, file));
+                file = path.join(file, objFile);
+              }
               win.webContents.send('open', file);
             },
           })),
